// src/types/pagination.ts
export interface LaravelPage<T> {
    current_page: number
    data: T[]
    first_page_url: string
    from: number | null
    last_page: number
    last_page_url: string
    links: { url: string | null; label: string; active: boolean }[]
    next_page_url: string | null
    path: string
    per_page: number
    prev_page_url: string | null
    to: number | null
    total: number
    total_roles: {
        all: RoleStats
        filtered: RoleStats
    }
}

export interface RoleStats {
    admin: 0
    teacher: 0
    parent: 0
}